+++ proton-mail-uniffi/src/mail/messages.rs
@@ -65,6 +65,9 @@ pub struct BodyOutput {
 
 #[uniffi::export]
 impl DecryptedMessage {
+    /// Gets the message body as an HTML. This does all of the transformations that are
+    /// required based on the options and the user settings.
+    ///
     /// # Parameters
     ///
     /// * `opts`: Which transform to apply to the html.
@@ -121,6 +124,60 @@ impl DecryptedMessage {
     pub fn mime_type(&self) -> MimeType {
         self.body.metadata.mime_type.into()
     }
+
+    #[must_use]
+    /// This is `Some` if the message is multipart. It contains the subject (if it has it) and the
+    /// attachments.
+    pub fn get_multipart_data(&self) -> Option<MultipartData> {
+        let attachments = self
+            .body
+            .pgp_attachments
+            .clone()?
+            .into_iter()
+            .map(|x| PgpAttachment {
+                id: x.id,
+                content_id: x.content_id,
+                name: x.name,
+                size: x.size as u64,
+                mime_type: x.mime_type,
+                data: x.data,
+            })
+            .collect_vec();
+
+        let subject = self.body.pgp_subject.clone();
+        Some(MultipartData {
+            subject,
+            attachments,
+        })
+    }
+}
+
+/// This comes from a multipart message, not to be confused with the other attachments.
+#[derive(Debug, PartialEq, Eq, Clone, Hash, uniffi::Record)]
+pub struct PgpAttachment {
+    /// Unique id across all attachments in an inbox.
+    pub id: String,
+    /// Content id extracted from mime.
+    pub content_id: String,
+    /// File name of the attachment.
+    pub name: String,
+    /// The size of the attachment in bytes.
+    pub size: u64,
+    /// The content type of the attachment.
+    ///
+    /// Is an empty string if no content type was found.
+    pub mime_type: String,
+    /// The attachment data.
+    pub data: Vec<u8>,
+}
+
+/// The extra data of a multipart message.
+#[derive(Debug, PartialEq, Eq, Clone, Hash, uniffi::Record)]
+pub struct MultipartData {
+    /// The subject that comes from a multipart message.
+    subject: Option<String>,
+    /// Attachments that come from a multipart message.
+    attachments: Vec<PgpAttachment>,
 }
 
 /// Get a specified message.